Fix vectormode to enum

+++ tt-train/sources/ttml/metal/ops/sdpa_bw/device/kernels/compute/sdpa_bw_compute_utils.hpp
@@ -68,7 +68,7 @@ inline void exp_full_tile(uint32_t idst) {
             /*CLAMP_NEGATIVE*/ false,
             DST_ACCUM_MODE>,
         idst,
-        (int)VectorMode::RC);
+        VectorMode::RC);
 #endif
 }
